Time: 15 ms (75.78%), Space: 199.1 MB (56.42%) - LeetHub

+++ 3834-merge-adjacent-equal-elements/3834-merge-adjacent-equal-elements.cpp
@@ -3,7 +3,7 @@ public:
     vector<long long> mergeAdjacent(vector<int>& nums) {
         vector<long long int>v;
         for(int i=0;i<nums.size();i++){
-            int curr= nums[i];
+            long  long int curr= nums[i];
             while(!v.empty() && v.back()==curr){
                 curr *=2;
                 v.pop_back();
